fix: address PR #213 review findings — clean up imports and test patterns
- Remove deferred `from pathlib import Path` and `from autoskillit.core
  import PIPELINE_FORBIDDEN_TOOLS` inside `_build_open_kitchen_prompt()`;
  move both to module level in `_prompts.py`
- Move `_OPEN_KITCHEN_CHOICE` and `_resolve_recipe_input` from
  `_init_helpers.py` to `_prompts.py` (they belong to the cook/picker
  workflow, not project-init)
- Update all import sites: `cli/__init__.py` and `cli/app.py` now import
  these symbols from `_prompts`, and `test_cook.py` import line updated
- Drop now-unused `TYPE_CHECKING`/`RecipeInfo` block from `_init_helpers.py`
- Replace fragile `_sys.modules.get(...) or importlib.import_module(...)`
  + `setattr(_app_mod, ...)` in two test_cook tests with the unambiguous
  `sys.modules["autoskillit.cli.app"]` pattern; avoids the `app` App-instance
  attribute shadow that makes string-based `setattr` fail
