+++ src/autoskillit/cli/__init__.py
@@ -10,11 +10,7 @@ import subprocess  # noqa: F401 — tests patch autoskillit.cli.subprocess.run
 from pathlib import Path  # noqa: F401 — tests patch autoskillit.cli.Path.home
 
 from autoskillit.cli._hooks import _claude_settings_path
-from autoskillit.cli._init_helpers import (
-    _OPEN_KITCHEN_CHOICE,
-    _prompt_recipe_choice,
-    _resolve_recipe_input,
-)
+from autoskillit.cli._init_helpers import _prompt_recipe_choice
 from autoskillit.cli._marketplace import (
     _clear_plugin_cache,
     _ensure_marketplace,
@@ -22,7 +18,12 @@ from autoskillit.cli._marketplace import (
     install,
     upgrade,
 )
-from autoskillit.cli._prompts import _build_open_kitchen_prompt, _build_orchestrator_prompt
+from autoskillit.cli._prompts import (
+    _OPEN_KITCHEN_CHOICE,
+    _build_open_kitchen_prompt,
+    _build_orchestrator_prompt,
+    _resolve_recipe_input,
+)
 from autoskillit.cli.app import (
     _generate_config_yaml,
     _prompt_test_command,
